fix: policies instead authorize method

+++ app/Http/Controllers/Account/DestroyController.php
@@ -9,6 +9,7 @@ class DestroyController extends BaseController
 {
     public function __invoke(Account $account): JsonResponse
     {
+        $this->authorize('delete', Account::class);
         $this->service->destroy($account);
 
         return response()->json([
